% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supportFunc.R
\name{check_arguments}
\alias{check_arguments}
\title{Test if user-supplied inputs match the expected arguments type and form}
\usage{
check_arguments(x, y)
}
\arguments{
\item{x}{The input provided by the user for the x argument.}

\item{y}{The input provided by the user for the y argument.}
}
\value{
There is no return value. If any error is found, the execution is just halted.
}
\description{
This function tests whether the input supplied by the user are of the expected type (class),
and are in the right form (overlap of extents and same projection).
}
