% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supportFunc.R
\name{raster_conversion}
\alias{raster_conversion}
\title{Convert the input raster object into terra's SpatRaster object}
\usage{
raster_conversion(y)
}
\arguments{
\item{y}{The input provided by the user for the y argument.}
}
\value{
the converted SpatRaster object
}
\description{
This function tests whether the supplied vector input is a simple features (sf) object,
if not, the object is converted using the st_as-sf() function.
}
