% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supportFunc.R
\name{Vector_conversion}
\alias{Vector_conversion}
\title{Convert the input vector object into simple features if it's provided as a Spatial object}
\usage{
Vector_conversion(x)
}
\arguments{
\item{x}{The input provided by the user for the x argument.}
}
\value{
the converted simple features object
}
\description{
This function tests whether the supplied vector input is a simple features (sf) object,
if not, the object is converted using the st_as-sf() function.
}
