% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Yanze_DEM}
\alias{Yanze_DEM}
\alias{DEM}
\title{Digital Elevation Model (DEM) of Yanze Watershed}
\format{
An S4 object of class stars
}
\usage{
DEM
}
\description{
This data consists of a 30X30m-resolution digital elevation model (DEM) of Yanze, a small watershed located in north-west of Kigali City, Rwanda.
The spatial reference system  used by the coordinates of DEM is GCS_WGS_1984 with a D_WGS_1984 datum.
The cell values are expressed in Meters with a minimum 1370m  of and a maximum of 2241.
}
\details{
extent: 29.91865, 30.03865, -1.941896, -1.755507  (xmin, xmax, ymin, ymax)
    dimensions : 671, 432, 289872  (nrow, ncol, ncell)
}
\keyword{DEM}
