% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Yanze_LULC}
\alias{Yanze_LULC}
\alias{lulcYanze}
\title{Land Use - Land Cover map of Yanze}
\format{
An S4 object of class stars
}
\usage{
lulcYanze
}
\description{
This data consists of a raster representing the landuse in Yanze watershed,
a small watershed located in north-west of Kigali City, Rwanda.
he spatial reference system  used by the coordinates of DEM is the WGS 84 UTM zone 35S - EPSG:32735.
Each land use/land cover class is represented by its code as follows:
1 = Forest
2 = Open Areas or Grass
3 = Agriculture (Seasonal)
5 = Settlements and Buildings
6 = Water
10 = Sparse Forest
11 = Agriculture (Perennial)
}
\details{
extent: 824728, 838128, -214942.7, -194282.7 (xmin, xmax, ymin, ymax)
}
\keyword{datasets}
