% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Yanze_sub-catchments-table}
\alias{Yanze_sub-catchments-table}
\alias{watersheds_df}
\title{Yanze sub-catchments table}
\format{
a data frame.
}
\usage{
watersheds_df
}
\description{
This data consists of a table containing the attributes of sub-catchment of Yanze watersehd.
}
\details{
Each rows represents a sub-catchment in yanze watershed, namely: Mulindi, cyonyonyo, and Yanze downstream.
\describe{
  \item{gridcode}{The name of each sub-catchment}
}
}
\keyword{datasets}
