% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Yanze_sub-catchments}
\alias{Yanze_sub-catchments}
\alias{watersheds}
\title{Yanze sub-catchments boundaries}
\format{
An S4 object of class simple features
}
\usage{
watersheds
}
\description{
This data consists of the boundaries of sub-catchments of Yanze watershed, a small watershed located in north-west of Kigali City, Rwanda.
The spatial reference system  used by the coordinates of this SpatialPolygonsDataFrame is GCS_WGS_1984 with a D_WGS_1984 datum.
}
\details{
extent: 29.91865, 30.03858, -1.755426, -1.941896 (xmin, xmax, ymin, ymax)
\describe{
  \item{gridcode}{The id of each polygon}
  \item{gridcode}{The name of each polygon}
}
}
\keyword{datasets}
